test(e2e): shared deck-fit Playwright harness + registry (closes #975)
UI-e2e P1 (generalizes #939). Lift the duplicated per-deck Playwright assertions
into test/playwright/harness/deck-fit-harness.mjs: settle barrier, mobile-fit
(per-element right<=innerWidth+1, no overflow-x exemption, vertical-clip),
section-id presence, CSP-meta guard, guard-the-guard; defineDeckSuite(entry)
generates the 3-test suite over DECK_REGISTRY. intro-deck.spec (187->10) and
deep-dive-deck.spec (200->10) become thin data registrations. Viewer dup
boilerplate factored into inspect-run-viewer-harness.mjs (it's an interactive
dashboard, not a fit-checked deck, so it keeps only the applicable section-id
guard). 13 Playwright tests pass; no coverage regression. Current artifacts only
(applied/observability decks were removed in #978/#985).
